requirements.txt
# ViralVortex — Python Dependencies
# Install with: pip install -r requirements.txt

# Core
anthropic>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0

# Trends
pytrends>=4.9.0

# Voice (fallback if ElevenLabs unavailable)
gTTS>=2.4.0

# Video / Image
Pillow>=10.0.0
# FFmpeg must be installed separately:
# Ubuntu/Debian: sudo apt install ffmpeg
# macOS: brew install ffmpeg

# YouTube Upload
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Thumbnail (optional DALL-E)
openai>=1.10.0

# Scheduler
schedule>=1.2.0
